Move paas metrics aiven client into own pkg
